[TASK] move convertAlphabetColumnToNumber to MainUtility so it can be used from other data adapters
