+++ Classes/Adapter/AbstractDefaultAdapter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Pixelant\PxaPmImporter\Adapter;
 
 use Pixelant\PxaPmImporter\Exception\InvalidAdapterFieldMapping;
+use Pixelant\PxaPmImporter\Utility\MainUtility;
 
 /**
  * Class AbstractDefaultAdapter
@@ -90,7 +91,7 @@ abstract class AbstractDefaultAdapter implements AdapterInterface
             if (is_numeric($configuration['mapping']['id'])) {
                 $this->identifier = (int)$configuration['mapping']['id'];
             } else {
-                $this->identifier = $this->convertAlphabetColumnToNumber($configuration['mapping']['id']);
+                $this->identifier = MainUtility::convertAlphabetColumnToNumber($configuration['mapping']['id']);
             }
         } else {
             throw new \RuntimeException('Adapter mapping require "id" (identifier) mapping to be set.', 1536050717594);
@@ -102,7 +103,7 @@ abstract class AbstractDefaultAdapter implements AdapterInterface
             foreach ($this->languagesMapping as $language => $languageMapping) {
                 foreach ($languageMapping as $field => $column) {
                     if (!is_numeric($column)) {
-                        $columnNumber = $this->convertAlphabetColumnToNumber($column);
+                        $columnNumber = MainUtility::convertAlphabetColumnToNumber($column);
                         $this->languagesMapping[$language][$field] = $columnNumber;
                     }
                 }
@@ -130,35 +131,7 @@ abstract class AbstractDefaultAdapter implements AdapterInterface
         throw new InvalidAdapterFieldMapping('Data column "' . $column . '" is not set', 1536051927592);
     }
 
-    /**
-     * Convert A to 0, B to 1 and so on
-     *
-     * @param string $column
-     * @return int
-     */
-    public function convertAlphabetColumnToNumber(string $column): int
-    {
-        /// @codingStandardsIgnoreStart
-        $alphabet = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'];
-        // @codingStandardsIgnoreEnd
 
-        $column = trim($column);
-
-        if (empty($column)) {
-            throw new \UnexpectedValueException('Column value could not be empty', 1536221838124);
-        }
-        $length = strlen($column);
-        if ($length > 2) {
-            throw new \LengthException('Maximum column value can be 2 chars', 1536221841673);
-        }
-
-        if ($length === 1) {
-            return array_search(strtoupper($column), $alphabet);
-        } else {
-            $firstValue = (array_search(strtoupper($column[0]), $alphabet) + 1) * count($alphabet);
-            return $firstValue + array_search(strtoupper($column[1]), $alphabet);
-        }
-    }
 
     /**
      * Convert source data according to mapping
