+++ Tests/Unit/Adapter/AbstractDefaultAdapter.php
@@ -166,27 +166,4 @@ class AbstractDefaultAdapterTest extends UnitTestCase
 
         $this->assertEquals($expect, $this->subject->_call('getFieldData', 1, $row));
     }
-
-    /**
-     * Convert excel columns A to 0, B to 1 and so on
-     * @test
-     */
-    public function convertAlphabetColumnToNumber()
-    {
-        $columnToExpect = [
-            'A' => 0,
-            'b' => 1,
-            'Aa' => 26,
-            'AB' => 27,
-            'AZ' => 51,
-            'SZ' => 519,
-        ];
-
-        foreach ($columnToExpect as $column => $expect) {
-            $this->assertEquals(
-                $expect,
-                $this->subject->_call('convertAlphabetColumnToNumber', $column)
-            );
-        }
-    }
 }
